vfio: move vfio backend into iommu/vfio
All of the container/group stuff belongs in iommu/ as well, leaving only
generic vfio stuff like irq and pci configuration in vfio/.
