+++ src/vfio/pci.c
@@ -46,7 +46,7 @@
 #include "ccan/minmax/minmax.h"
 #include "ccan/str/str.h"
 
-#include "container.h"
+#include "iommu/container.h"
 
 static int pci_set_bus_master(struct vfio_pci_device *pci)
 {
